pyproject.toml
# pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "blr-aqi-forecasting"
version = "0.1.0"
description = "Production-grade AQI forecasting & health-aware routing system for Bangalore."
authors = ["AI/ML Engineer"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
pandas = "^2.0.3"
numpy = "^1.24.3"
scikit-learn = "^1.3.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
pyyaml = "^6.0.1"
optuna = "^3.3.0"
mlflow = "^2.6.0"
networkx = "^3.1"
osmnx = "^1.6.0"
geopandas = "^0.13.2"
redis = "^5.0.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.5.1"
ruff = "^0.0.285"
black = "^23.7.0"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 100
